requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
django-redis==5.4.0
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
types-python-dateutil==2.9.0.20240906
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.7.0
